requirements.txt
asgiref==3.8.1
decorator==5.1.1
Django==5.1.4
djangorestframework==3.15.2
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
future==1.0.0
imageio==2.36.1
imageio-ffmpeg==0.5.1
moviepy==2.1.1
numpy==2.2.0
pillow==10.4.0
proglog==0.1.10
python-dotenv==1.0.1
setuptools==75.6.0
sqlparse==0.5.3
tqdm==4.67.1
